V0.3.1.1.1230 Beta 完成linux支持，完善部分frpc信息翻译功能
V 3.1.1.1230 Beta
更新内容：
1.完成对linux支持
2.修复系统判断的bug
3.新增启动系统类型提示
4.完善部分frpc信息内容

+++ MossFrpProcess/src/main/java/org/mossmc/mosscg/MossFrpProcess.java
@@ -32,11 +32,30 @@ public class MossFrpProcess {
 
     //检查是否用参数启动
     //避免有用户直接点软件导致出bug
+    //以及初始化系统设定
     public static void checkStart(String args) {
         if (!args.contains("-MossFrp=nb")) {
             sendInfo("please not start this!");
             System.exit(1);
         }
+        if (args.contains("-systemType=linux")) {
+            getSystemType = systemType.linux;
+        }
+        if (args.contains("-systemType=windows")) {
+            getSystemType = systemType.windows;
+        }
+        if (getSystemType == null) {
+            sendInfo("please not start this!");
+            System.exit(1);
+        }
+    }
+
+    //系统类型
+    //没加参数就默认win
+    public static systemType getSystemType;
+
+    public enum systemType {
+        windows,linux
     }
 
     //读取请求
@@ -100,16 +119,26 @@ public class MossFrpProcess {
         sendInfo("debug frp run");
         Runtime run = Runtime.getRuntime();
         try {
-            sendInfo("send Start "+path);
-            Process frp;
-            frp = run.exec("taskkill /im frpc-"+path+".exe /f");
-            BufferedReader output = new BufferedReader(new InputStreamReader(frp.getInputStream()));
-            output.readLine();
-            output.close();
-            frp = run.exec("./MossFrp/frps/"+path+"/frpc-"+path+".exe -c ./MossFrp/frps/"+path+"/frpc.ini");
-            frpProcessMap.put(path,frp);
-            sendInfo("send Success "+path);
-            readFrp(path);
+            if (getSystemType == systemType.windows) {
+                sendInfo("send Start "+path);
+                Process frp;
+                frp = run.exec("taskkill /im frpc-"+path+".exe /f");
+                BufferedReader output = new BufferedReader(new InputStreamReader(frp.getInputStream()));
+                output.readLine();
+                output.close();
+                frp = run.exec("./MossFrp/frps/"+path+"/frpc-"+path+".exe -c ./MossFrp/frps/"+path+"/frpc.ini");
+                frpProcessMap.put(path,frp);
+                sendInfo("send Success "+path);
+                readFrp(path);
+            }
+            if (getSystemType == systemType.linux) {
+                sendInfo("send Start "+path);
+                Process frp;
+                frp = run.exec("./MossFrp/frps/"+path+"/frpc-"+path+" -c ./MossFrp/frps/"+path+"/frpc.ini");
+                frpProcessMap.put(path,frp);
+                sendInfo("send Success "+path);
+                readFrp(path);
+            }
         }catch (Exception e){
             e.printStackTrace();
             sendInfo("send Failed "+path);
